refactor: 커스텀 에러 제거 및 에러 출력문 상수화
