+++ src/validator/retry.js
@@ -1,8 +1,7 @@
-import { GAME } from '../constants/lottoConstants.js';
-import { CustomError, ERROR_CODE } from '../utils/Error.js';
+import { GAME, ERROR } from '../constants/constants.js';
 
 export default function checkRetryFormat(command) {
   if (command !== GAME.RETRY && command !== GAME.EXIT) {
-    throw new CustomError({ code: ERROR_CODE.INVALID_FORMAT }, command);
+    throw new Error(ERROR.INVALID_FORMAT);
   }
 }
